+++ truffle/CHANGELOG.md
@@ -19,6 +19,8 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Added [Breakpoint.Kind](http://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/debug/Breakpoint.Kind.html) to distinguish different breakpoint kinds.
 * Added [ResultVerifier.getDefaultResultVerifier](http://www.graalvm.org/truffle/javadoc/org/graalvm/polyglot/tck/ResultVerifier.html#getDefaultResultVerifier--).
 * Added [addToHostClassPath](http://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/TruffleLanguage.Env.html#getPolyglotBindings--) method that can be used to allow guest language users to add to the host class path.
+* Added new permission TruffleLanguage.Env#isNativeAccessAllowed to control access to the Truffle NFI.
+* Changed default permissions in language launchers to full access. The embedding API still defaults to restricted access.
 
 ## Version 0.33